requirements.txt
pyyaml>=6.0
lxml>=4.9
jinja2>=3.1
requests>=2.31
openpyxl>=3.1
pytest>=7.4
rich>=13.0
click>=8.1
streamlit>=1.30
plotly>=5.18
pandas>=2.0
